refactor: remove `bttf.modules` package;
