--- pypots/forecasting/bttf/modules/__init__.py
@@ -1,12 +1,0 @@
-"""
-
-"""
-
-# Created by Wenjie Du <wenjay.du@gmail.com>
-# License: BSD-3-Clause
-
-from .core import BTTF_forecast
-
-__all__ = [
-    "BTTF_forecast",
-]
